requirements.txt
PyAutoGUI~=0.9.54
requests~=2.32.5
pillow~=12.1.1
anthropic~=0.84.0
python-dotenv~=1.2.2
pyperclip~=1.11.0
sentence-transformers>=3.0.0
numpy>=1.26.0